LayoutTests/imported/w3c/web-platform-tests/html/anonymous-iframe/embedding.tentative.https.window-expected.txt


PASS Parent embeds same-origin credentialless iframe
PASS Parent embeds cross-origin credentialless iframe
FAIL COEP:require-corp parent embeds same-origin credentialless iframe assert_equals: expected "load" but got "block"
FAIL COEP:require-corp parent embeds cross-origin credentialless iframe assert_equals: expected "load" but got "block"
PASS COEP:credentialless parent embeds same-origin credentialless iframe
PASS COEP:credentialless parent embeds cross-origin credentialless iframe
FAIL COOP:same-origin + COEP:require-corp embeds same-origin credentialless iframe assert_equals: expected "load" but got "block"
FAIL COOP:same-origin + COEP:require-corp embeds cross-origin credentialless iframe assert_equals: expected "load" but got "block"
PASS COOP:same-origin + COEP:credentialless embeds same-origin credentialless iframe
PASS COOP:same-origin + COEP:credentialless embeds cross-origin credentialless iframe
PASS Parents embeds a CSP:frame-ancestors credentialless iframe
PASS Cross-Origin-Isolated parent embeds same-origin COEP credentialless iframe
FAIL Cross-Origin-Isolated parent embeds cross-origin COEP credentialless iframe assert_equals: expected "load" but got "block"
